requirements.txt
python-telegram-bot==20.3
Flask
nest_asyncio
g4f
